pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alpha-fpl"
version = "0.1.0"
description = "Risk-Aware Portfolio Optimization Engine for Fantasy Premier League"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Alpha-FPL Team"},
]
keywords = ["fpl", "fantasy-football", "bayesian", "optimization", "copula"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core data processing
    "polars>=0.20.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pyarrow>=14.0.0",
    
    # Data ingestion
    "soccerdata>=1.6.0",
    "understatapi>=0.1.0",
    "thefuzz[speedup]>=0.20.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    
    # Bayesian inference (JAX/NumPyro)
    "jax>=0.4.20",
    "jaxlib>=0.4.20",
    "numpyro>=0.13.0",
    
    # Optimization
    "cvxpy>=1.4.0",
    "gurobipy>=11.0.0",
    
    # Configuration & tracking
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    
    # Utilities
    "tqdm>=4.66.0",
    "rich>=13.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
]

colab = [
    # TPU/GPU specific JAX
    "jax[tpu]>=0.4.20",
]

[project.scripts]
alpha-fpl = "src.run_pipeline:main"
generate-id-map = "tools.generate_map:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
